+++ android/app/src/main/java/com/now/naaga/data/remote/retrofit/ServicePool.kt
@@ -1,7 +1,6 @@
 package com.now.naaga.data.remote.retrofit
 
 import com.now.naaga.data.remote.retrofit.RetrofitFactory.retrofit
-import com.now.naaga.data.remote.retrofit.service.Adventure2Service
 import com.now.naaga.data.remote.retrofit.service.AdventureService
 import com.now.naaga.data.remote.retrofit.service.PlaceService
 import com.now.naaga.data.remote.retrofit.service.RankService
@@ -9,7 +8,6 @@ import com.now.naaga.data.remote.retrofit.service.StatisticsService
 
 object ServicePool {
     val adventureService = retrofit.create(AdventureService::class.java)
-    val adventureService2 = retrofit.create(Adventure2Service::class.java)
     val rankService = retrofit.create(RankService::class.java)
     val statisticsService = retrofit.create(StatisticsService::class.java)
     val placeService = retrofit.create(PlaceService::class.java)
