+++ android/app/src/main/java/com/now/naaga/data/remote/retrofit/service/StatisticsService.kt
@@ -1,6 +1,6 @@
 package com.now.naaga.data.remote.retrofit.service
 
-import com.now.naaga.data.remote.dto.third.StatisticsDto
+import com.now.naaga.data.remote.dto.StatisticsDto
 import retrofit2.Call
 import retrofit2.http.GET
